[nemu]: fix function sp align 128 bits

+++ navy-apps/libs/libos/src/crt0/crt0.c
@@ -5,8 +5,8 @@
 int main(int argc, char *argv[], char *envp[]);
 extern char **environ;
 void call_main(uintptr_t *args) {
-  int argc = *(int *)args;
-  char **argv = (char **)((unsigned char *)args + sizeof(int));
+  int argc = *(uintptr_t *)args;
+  char **argv = (char **)((unsigned char *)args + sizeof(uintptr_t));
   char **envp = argv + (argc + 1);
   exit(main(argc, argv, envp));
   assert(0);
